pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nexus-client"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=1.6.0",
    "aiohttp>=3.8.0",
]
